new_corner electron event triggers event in watcher
*  moved cycling_flag and button_callback() to drawer.py
*  electron-server translation event type for corners is new_corner
